Restructure pixl core
Make modules more consistent
