+++ cli/tests/conftest.py
@@ -17,7 +17,7 @@ from __future__ import annotations
 import pathlib
 
 import pytest
-from core.database import Base, Extract, Image
+from core.db.models import Base, Extract, Image
 from sqlalchemy import Engine, create_engine
 from sqlalchemy.orm import Session, sessionmaker
