+++ pixl_core/src/core/db/__init__.py
@@ -1,0 +1,1 @@
+"""Define PIXL pipeline database and functionality."""
