+++ pixl_core/src/core/rest_api/__init__.py
@@ -1,0 +1,1 @@
+"""Shared REST API definitions."""
