+++ pixl_core/tests/test_upload.py
@@ -17,7 +17,7 @@
 from datetime import datetime, timezone
 
 import pytest
-from core.database import Image
+from core.db.models import Image
 from core.queries import get_project_slug_from_db, update_exported_at
 from core.upload import upload_dicom_image
